refactor(chat): extract ComposerView from ChatView
Extract the composer area (text field, action buttons, attachment strip,
scroll indicator, microphone button, scroll offset reader) into a
standalone ComposerView for reuse in the upcoming workspace layout.
